Add connection types, and multiple hostname input ability
