+++ src/views/webview-app/components/connect-form/advanced-tab/advanced-tab.tsx
@@ -1,13 +1,11 @@
 import * as React from 'react';
 
 import ReadPreferenceSelect from './read-preference-select';
-import ReplicaSetInput from './replica-set-input';
 
 class AdvancedTab extends React.Component {
   render(): React.ReactNode {
     return (
       <React.Fragment>
-        <ReplicaSetInput />
         <ReadPreferenceSelect />
       </React.Fragment>
     );
