+++ src/views/webview-app/components/connect-form/connection-form.less
@@ -53,7 +53,7 @@
       right: -1px;
       pointer-events: none;
       outline: none;
-      background-color: var(--vscode-focusBorder);
+      background-color: #C3E7CA;
       border-top-left-radius: 3px;
       border-top-right-radius: 3px;
       animation-name: selectedConnectionTabAnimation;
