+++ src/views/webview-app/components/connect-form/general-tab/host/host.less
@@ -1,0 +1,54 @@
+
+.host-input-area {
+  margin: 0;
+  margin-top: 15px;
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+  align-items: center;
+
+  &:first-child {
+    margin-top: 0;
+  }
+}
+
+.host-input-port {
+  margin: 0;
+  max-width: 30%;
+  margin-left: 10px;
+}
+
+.host-input-host {
+  margin: 0;
+  margin-right: 10px;
+}
+
+.host-add-host-button {
+  display: inline-block;
+  margin: 0;
+  margin-left: 10px;
+  margin-top: 22px;
+  padding: 5px;
+  background: none;
+  color: var(--vscode-editor-foreground);
+  border: none;
+
+  &:hover {
+    cursor: pointer;
+  }
+}
+
+.host-remove-host-button {
+  display: inline-block;
+  margin: 0;
+  margin-left: 10px;
+  margin-top: 22px;
+  padding: 5px;
+  background: none;
+  color: var(--vscode-editor-foreground);
+  border: none;
+
+  &:hover {
+    cursor: pointer;
+  }
+}
